+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -83,6 +83,129 @@ class ConfigurationTest extends TestCase
         ];
     }
 
+    #[DataProvider('provideEquivalentProfilerExclusions')]
+    public function testProfilerExcludedHttpCodesAreNormalized(array $excludedHttpCodes)
+    {
+        $config = (new Processor())->processConfiguration(new Configuration(true), [[
+            'profiler' => ['excluded_http_codes' => $excludedHttpCodes],
+        ]]);
+
+        $this->assertSame([
+            404 => [],
+            400 => ['^/foo', '^/bar'],
+        ], $config['profiler']['excluded_http_codes']);
+    }
+
+    public static function provideEquivalentProfilerExclusions(): iterable
+    {
+        yield 'mapping' => [[404 => null, 400 => ['^/foo', '^/bar']]];
+        yield 'mapping with an explicit empty list' => [[404 => [], 400 => ['^/foo', '^/bar']]];
+        yield 'mapping with true' => [[404 => true, 400 => ['^/foo', '^/bar']]];
+        yield 'mixed forms' => [[404, 400 => ['^/foo', '^/bar']]];
+    }
+
+    public function testProfilerExcludedHttpCodesAcceptsABareStatusCode()
+    {
+        $config = (new Processor())->processConfiguration(new Configuration(true), [[
+            'profiler' => ['excluded_http_codes' => 404],
+        ]]);
+
+        $this->assertSame([404 => []], $config['profiler']['excluded_http_codes']);
+    }
+
+    public function testProfilerExcludedHttpCodesAcceptsASinglePathAsAString()
+    {
+        $config = (new Processor())->processConfiguration(new Configuration(true), [[
+            'profiler' => ['excluded_http_codes' => [404 => '^/foo']],
+        ]]);
+
+        $this->assertSame([404 => ['^/foo']], $config['profiler']['excluded_http_codes']);
+    }
+
+    public function testProfilerExcludedHttpCodesSkipsCodesDisabledWithFalse()
+    {
+        $config = (new Processor())->processConfiguration(new Configuration(true), [[
+            'profiler' => ['excluded_http_codes' => [404 => false, 400 => null]],
+        ]]);
+
+        $this->assertSame([400 => []], $config['profiler']['excluded_http_codes']);
+    }
+
+    public function testProfilerExcludedHttpCodesAreOverriddenAcrossFiles()
+    {
+        $config = (new Processor())->processConfiguration(new Configuration(true), [
+            ['profiler' => ['excluded_http_codes' => [404]]],
+            ['profiler' => ['excluded_http_codes' => [404 => ['^/api']]]],
+        ]);
+
+        $this->assertSame([404 => ['^/api']], $config['profiler']['excluded_http_codes']);
+    }
+
+    public function testProfilerExcludedPathsAcceptsABareRegularExpression()
+    {
+        $config = (new Processor())->processConfiguration(new Configuration(true), [[
+            'profiler' => ['excluded_paths' => '^/\.well-known/'],
+        ]]);
+
+        $this->assertSame(['^/\.well-known/'], $config['profiler']['excluded_paths']);
+    }
+
+    #[DataProvider('provideOutOfRangeHttpCodes')]
+    public function testProfilerExcludedHttpCodesRejectsCodesOutOfRange(int $code, string $message)
+    {
+        $this->expectException(InvalidConfigurationException::class);
+        $this->expectExceptionMessage($message);
+
+        (new Processor())->processConfiguration(new Configuration(true), [[
+            'profiler' => ['excluded_http_codes' => [$code]],
+        ]]);
+    }
+
+    public static function provideOutOfRangeHttpCodes(): iterable
+    {
+        yield 'below the lower bound' => [99, 'only accepts HTTP status codes between 100 and 599, "99" given'];
+        yield 'above the upper bound' => [600, 'only accepts HTTP status codes between 100 and 599, "600" given'];
+        yield 'negative' => [-1, 'only accepts HTTP status codes between 100 and 599, "-1" given'];
+    }
+
+    public function testProfilerExcludedPathsRejectsEmptyPatterns()
+    {
+        $this->expectException(InvalidConfigurationException::class);
+
+        (new Processor())->processConfiguration(new Configuration(true), [[
+            'profiler' => ['excluded_paths' => ['']],
+        ]]);
+    }
+
+    public function testProfilerExcludedHttpCodesRejectsEmptyPathPatterns()
+    {
+        $this->expectException(InvalidConfigurationException::class);
+
+        (new Processor())->processConfiguration(new Configuration(true), [[
+            'profiler' => ['excluded_http_codes' => [404 => ['']]],
+        ]]);
+    }
+
+    public function testProfilerExcludedPathsRejectsInvalidRegularExpressions()
+    {
+        $this->expectException(InvalidConfigurationException::class);
+        $this->expectExceptionMessage('Invalid regular expression in the "excluded_paths" option');
+
+        (new Processor())->processConfiguration(new Configuration(true), [[
+            'profiler' => ['excluded_paths' => ['^/foo(']],
+        ]]);
+    }
+
+    public function testProfilerExcludedHttpCodesRejectsInvalidRegularExpressions()
+    {
+        $this->expectException(InvalidConfigurationException::class);
+        $this->expectExceptionMessage('Invalid regular expression in the "excluded_http_codes" option');
+
+        (new Processor())->processConfiguration(new Configuration(true), [[
+            'profiler' => ['excluded_http_codes' => [404 => ['^/foo(']]],
+        ]]);
+    }
+
     public function testCacheAppAndDefaultProviderCannotBeCombined()
     {
         $this->expectException(InvalidConfigurationException::class);
@@ -960,6 +1083,8 @@ class ConfigurationTest extends TestCase
                 'enabled' => false,
                 'only_exceptions' => false,
                 'only_main_requests' => false,
+                'excluded_paths' => [],
+                'excluded_http_codes' => [],
                 'dsn' => 'file:%kernel.cache_dir%/profiler',
                 'collect' => true,
                 'collect_parameter' => null,