+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/profiler_exclusions.php
@@ -1,0 +1,12 @@
+<?php
+
+$container->loadFromExtension('framework', [
+    'profiler' => [
+        'enabled' => true,
+        'excluded_paths' => ['^/\.well-known/'],
+        'excluded_http_codes' => [
+            404 => null,
+            400 => ['^/foo', '^/bar'],
+        ],
+    ],
+]);